app/api/v1/endpoints/refunds.py
from fastapi import APIRouter, Depends, HTTPException, status
from sqlalchemy.orm import Session
from typing import Any, Dict
import logging

from app.db.session import get_db
from app.schemas.refund import RefundRequestBase, RefundResponse
from app.crud import refund as crud_refund

router = APIRouter()
logger = logging.getLogger(__name__)

@router.post(
    "/orders/{order_id}/refund",
    response_model=RefundResponse,
    status_code=status.HTTP_202_ACCEPTED,
    summary="Request a refund for an order",
    description="Stub endpoint that logs the refund request and returns a pending status"
)
async def request_refund(
    order_id: str,
    refund_request: RefundRequestBase,
    db: Session = Depends(get_db)
) -> Dict[str, str]:
    """
    Stub endpoint for processing refund requests.
    Logs the request and returns a pending status.
    """
    try:
        # Log the refund request
        logger.info(f"Processing refund request for order {order_id}")
        logger.debug(f"Refund request data: {refund_request.dict()}")
        
        # Create a refund request record in the database
        db_refund = crud_refund.create_refund_request(
            db=db,
            order_id=order_id,
            platform=refund_request.platform,
            request_data=refund_request.dict()
        )
        
        # Log the created refund request ID
        logger.info(f"Created refund request with ID: {db_refund.id}")
        
        # Return the pending status response
        return {"status": "pending"}
        
    except Exception as e:
        logger.error(f"Error processing refund request: {str(e)}", exc_info=True)
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="An error occurred while processing the refund request"
        )
