requirements-test.txt
pytest==7.4.2
pytest-cov==4.1.0
httpx==0.25.0
pytest-asyncio==0.21.1
sqlalchemy-utils==0.41.1
pytest-mock==3.11.1
